requirements.txt
streamlit>=1.48.0
ibm-watsonx-ai>=1.3.32
ibm-watson>=10.0.0
chromadb>=1.0.15
pypdf>=4.0.0
python-docx>=1.2.0
openpyxl>=3.1.0
pandas>=2.2.0
pytesseract>=0.3.13
Pillow>=10.0.0
pdf2image>=3.1.0
reportlab>=4.0.0
streamlit-extras>=0.7.5

